refactor: api fetchinig 로직과 이를 이용하는 custom hook을 분리
