+++ src/Apis/auth.ts
@@ -1,9 +1,5 @@
 import { API_END_POINT } from '@/Constants/api';
-import { useLoginStore } from '@/store/auth';
 import { httpClient } from '@/utils/axios';
-import { useMutation } from '@tanstack/react-query';
-
-import { useNavigate } from 'react-router-dom';
 
 export interface UserInfo {
   email: string;
@@ -11,24 +7,6 @@ export interface UserInfo {
 }
 
 export const logOut = async () => httpClient.post(API_END_POINT.LOGOUT);
-
-export const useLogOutMutation = () => {
-  const { setIsLoggedOut } = useLoginStore();
-  const navigate = useNavigate();
-  const { mutate } = useMutation({
-    mutationKey: ['logout'],
-    mutationFn: () => logOut(),
-    onSuccess: () => {
-      // console.log('로그아웃 성공');
-      setIsLoggedOut();
-      navigate('/');
-    },
-    onError: () => {
-      // console.log('로그아웃 실패');
-    },
-  });
-  return { mutate };
-};
 
 export const getUser = async () => {
   const response = await httpClient.get(API_END_POINT.USER);
