+++ src/Components/ApplicantCard/index.tsx
@@ -4,7 +4,8 @@ import styled from 'styled-components';
 import { InfoField } from '../Common/InfoField';
 import Button from '../Common/Button';
 import { useState } from 'react';
-import { useAcceptApplyMutation, useRefuseApplyMutation } from '@/Apis/study';
+import { useAcceptApplyMutation } from '@/Hooks/study/useAcceptApplyMutation';
+import { useRefuseApplyMutation } from '@/Hooks/study/useRefuseApplyMutation';
 import Modal from '../Common/Modal';
 import { APPLY } from '@/Constants/messages';
 import { useModalStore } from '@/store/modal';
