+++ src/Pages/RecruitmentDetail/index.tsx
@@ -2,7 +2,8 @@ import styled from 'styled-components';
 import { InfoField } from '../../Components/Common/InfoField';
 import { RowDivider } from '../../Components/Common/Divider/RowDivider';
 import { ColumnDivider } from '../../Components/Common/Divider/ColumnDivider';
-import { useCloseRecruitmentMutation, useRecruitmentDetail } from '@/Apis/recruitment';
+import { useRecruitmentDetail } from '@/Hooks/recruitments/useRecruitmentDetail';
+import { useCloseRecruitmentMutation } from '@/Hooks/recruitments/useCloseRecruitmentMutation';
 import { useNavigate, useParams } from 'react-router-dom';
 import { dateFormatter, getPeriod, isEdited } from '@/utils/date';
 import RecruitmentInfoSection from './RecruitmentInfoSection';
