+++ src/Pages/Studies/CreateRecruitment.tsx
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-// import { ContactButton } from '../../Components/Selectbox/ContactButton';
 import { ContactUrlInput } from '../../Components/Textarea/ContactUrlInput';
 import { SubmitButton } from '../../Components/Button/Studies/SubmitButton';
 import { CalendarButton } from '../../Components/Selectbox/CalendarButton';
@@ -7,23 +6,18 @@ import { PositionButton } from '../../Components/Selectbox/PositionButton';
 import { StackSelectButton } from '@/Components/Selectbox/StackSelectButton';
 import { Mainarea } from '../../Components/Textarea/Mainarea';
 import { Titlearea } from '../../Components/Textarea/Titlearea';
-// import { GatherButton } from '../../Components/Selectbox/GatherButton';
-// import { StackModal } from '../../Components/Modal/
 import { EndDate } from '../../Components/Calendar/EndDate';
 import { media } from '../../Styles/theme';
-// import { Gather } from '@/Types/studies';
 import { useState } from 'react';
-// import { stackCategory } from '@/Shared/category';
-import { useStack } from '@/Apis/stack';
+import { useStack } from '@/Hooks/stack/useStack';
 import { useNavigate, useParams } from 'react-router-dom';
 import axios from 'axios';
 import { One, Two, Three, Four } from '@/Assets';
 import { SaveButton } from '@/Components/Button/Studies/SaveButton';
-// import { STUDY } from '@/Constants/queryString';
 import { Gather } from '@/Types/studies';
 import { ContactButton } from '@/Components/Selectbox/ContactButton';
 import { ApplicantButton } from '@/Components/Selectbox/ApplicantButton';
-import { useStudyDetail } from '@/Apis/study';
+import { useStudyDetail } from '@/Hooks/study/useStudyDetail';
 
 axios.defaults.withCredentials = true;
 export type OptionalCreates = Partial<Gather>;
@@ -31,9 +25,6 @@ export type OptionalCreates = Partial<Gather>;
 const CreateRecruitment = () => {
   const studyId = Number(useParams().studyId);
 
-  // console.log(data);
-  // const { d } = useStudyId(studyId as any);
-  // console.log(d);
   const Navigation = useNavigate();
   // const studyId = Number(useParams().studyId);
   useStudyDetail(studyId);
