+++ src/Providers/LoginProvider/index.tsx
@@ -18,7 +18,6 @@ const LoginProvider = ({ children }: LoginProviderProps) => {
         setIsLoggedIn();
         setUser(data);
       } catch (e) {
-        // console.log(e);
         setIsLoggedOut();
         resetUser();
       }
