+++ src/utils/axios.ts
@@ -11,17 +11,14 @@ export const createClient = (config?: AxiosRequestConfig) => {
     ...config,
   });
   axiosInstance.interceptors.request.use((request) => {
-    // console.log(request);
     return request;
   });
 
   axiosInstance.interceptors.response.use(
     (response) => {
-      // console.log(response);
       return response;
     },
     (error) => {
-      // console.log(error);
       return Promise.reject(error);
     },
   );
